♻ Split one file server module into directory-based module
